+++ src/main/java/com/boopugstudios/dynamicjobseconomy/DynamicJobsEconomy.java
@@ -8,7 +8,6 @@ import com.boopugstudios.dynamicjobseconomy.integrations.IntegrationManager;
 import com.boopugstudios.dynamicjobseconomy.jobs.JobManager;
 import com.boopugstudios.dynamicjobseconomy.admin.AdminAuditLogger;
 import com.boopugstudios.dynamicjobseconomy.admin.AdminConfirmationManager;
-import com.boopugstudios.dynamicjobseconomy.admin.AdminReasonChatListener;
 import com.boopugstudios.dynamicjobseconomy.business.ConsolidatedBusinessManager;
 import com.boopugstudios.dynamicjobseconomy.gigs.GigManager;
 import com.boopugstudios.dynamicjobseconomy.notifications.NotificationManager;
@@ -218,7 +217,6 @@ public final class DynamicJobsEconomy extends JavaPlugin {
         getServer().getPluginManager().registerEvents(new JobListener(this), this);
         getServer().getPluginManager().registerEvents(new BusinessListener(this), this);
         getServer().getPluginManager().registerEvents(notificationManager, this);
-        getServer().getPluginManager().registerEvents(new AdminReasonChatListener(this), this);
     }
     
     // Getters for managers